Split mc_move_set.hpp into multiple header files and format them
